[dev8] Factor out captureLogs
